Fixes for Test App Ambiguities on SunSolaris
A few test apps were experiencing build issues on SunSolaris related to
ambiguities with certain cmath methods. Changes have been made to rectify
the current issues.

- Stats.hpp - An early fix cast the SQRT calls to always do long double
  and cast back. This cast has been removed.
- Stats_T.cpp - Has been revised in its entirety. Test now performs all
  methods of the Stats class for each of its expected floating point
  types.
- AlmOrbit_T.cpp - A variable was treated as a floating point value when it
  was a long int. This led to ambiguities with fabs during an assert comparison.
  The fabs has been removed and the comparison corrected to do an integer
  comparison.
- stl_helpers.hpp - The use of std::find did not have the std:: portion leading
  to an error when that particular method was called.
- Stl_helpers_T.cpp - Expected values for the floating point comparisons in
  certain assert lines were left as integers. However, those integers were
  placed inside of fabs. Upon promotion to floating point values, those
  integers would lead to the ambiguity error on SunSolaris compilers. These
  integers have been properly cast to the proper type first.

+++ core/lib/Math/Stats.hpp
@@ -103,7 +103,7 @@ namespace gpstk
       inline T StdDev(void) const
       {
          if(n <= 1) return T();
-         return SQRT( (long double) Variance() );
+         return SQRT( Variance() );
       }
 
       /// return the normalization constant = sum weights
@@ -401,10 +401,10 @@ namespace gpstk
       }
 
       /// return computed X standard deviation
-      inline T StdDevX(void) const { return SQRT( (long double) VarianceX() ); }
+      inline T StdDevX(void) const { return SQRT( VarianceX() ); }
 
       /// return computed Y standard deviation
-      inline T StdDevY(void) const { return SQRT( (long double) VarianceY() ); }
+      inline T StdDevY(void) const { return SQRT( VarianceY() ); }
 
       /// return slope of best-fit line Y=slope*X + intercept
       inline T Slope(void) const
